+++ docs/changelog.rst
@@ -8,6 +8,7 @@ Changelog
 Features
 ~~~~~~~~
 - `#8217 <https://0xacab.org/leap/bitmask-dev/issues/8217>`_: renew OpenPGP keys before they expire.
+- `#9074 <https://0xacab.org/leap/bitmask-dev/issues/9074>`_: pin provider ca certs.
 - Set a windows title, so that Bitmask windows can be programmatically manipulated.
 
 Misc